# Add simple benchmarks for osrm-* tools